% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrAssign.R
\name{CorrAssign1}
\alias{CorrAssign1}
\title{Assign label based on correlation result}
\usage{
CorrAssign1(cor.matrix, sample.label, quantile.cutoff)
}
\arguments{
\item{cor.matrix}{matrix of correlation where each column contains correlations for each samples. Row corresponds to sample in label.}

\item{sample.label}{specify asasy data in GEOdata to calculate pair-wise correlation, default is "exprs"}

\item{quantile.cutoff}{lower bound for quantile of Spearman correlation to be averaged by each label}
}
\value{
A list containing (1) label result (2) corAVG.Top.pct: average correlation of each sample by each label (3) corSD.Top.pct: sd of correlation of each sample by each label
}
\description{
Assign label based on correlation result by given a quantile of correlation.
}
\examples{

CorrAssign1(cor.matrix = spear.cor, sample.label = stage, quantile.cutoff = 0.5)

}
