% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_functions.R
\name{corrSpearman}
\alias{corrSpearman}
\title{Calculate Spearman correlation given two matrices}
\usage{
corrSpearman(x, y)
}
\arguments{
\item{x}{matrix of samples. (Row: gene; column: samples)}

\item{y}{matrix of samples. (Row: gene; column: samples)}
}
\value{
A matrix.
}
\description{
Calculate Spearman correlation given two matrices
}
