% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMGMM.R
\name{EMGMM}
\alias{EMGMM}
\title{Classification using Gaussian mixture model with EM algorithm.}
\usage{
EMGMM(expression, ref_mean, ref_sd, subtype, true_label)
}
\arguments{
\item{expression}{a gene expression matrix, row = G genes, column = N samples/cells}

\item{ref_mean}{mean each marker for each cluster, row = L markers (L<G), column = K clusters}

\item{ref_sd}{sd each marker for each cluster, row = L markers (L<G), column = K clusters}

\item{subtype}{a vector contains names of clusters}

\item{true_label}{a vector contains actual cluster label of each sample/cells}
}
\value{
A list of three items:

1: a matrix of probability of each sample belongs to each cluster

2: assignment based on maximum likelihood

3: EM algorithm iteration
}
\description{
This function classify samples/cells into different pre-specified groups,
where mean and sd for Gaussian distribution are given as reference/initial input.
This is a supervised clustering method.
}
