% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureFilter.R
\name{FeatureFilter}
\alias{FeatureFilter}
\title{Select features/genes/probes based on a pre-define vector.}
\usage{
FeatureFilter(GEOdata, filterVar = NA, filterValue, unique = FALSE)
}
\arguments{
\item{GEOdata}{either a matrix or SummerizedExperiment object contains expression.}

\item{filterVar}{a character for the variable that is being subsetted}

\item{filterValue}{a vector of genes/probes identities one wish to keep}

\item{unique}{a boolean, after filter, whether to keep only those}
}
\value{
A SummerizedExperiment object contains both expression and metadata related to each sample.

Or a matrix of expression data.
}
\description{
This function will subset expression data based on a pre-define vector of genes or probes.
}
\examples{

GeoExtract(GEOAccession = "GSE48762", getGPL = TRUE,  plots = TRUE, NSub = 20)

}
