% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOExtract.R
\name{GEOExtract}
\alias{GEOExtract}
\title{Extract data from GEO.}
\usage{
GEOExtract(
  GEOAccession,
  GSEMatrix = TRUE,
  getGPL = TRUE,
  plots = TRUE,
  NSub = 10
)
}
\arguments{
\item{GEOAccession}{character contains GEO accession}

\item{GSEMatrix}{boolean, indicating whether to extract the GSE matrix from GEO, with a default value of TRUE.}

\item{getGPL}{boolean, indicating whether to  extract GPL information or not from GEO, with a default value of TRUE.}

\item{plots}{boolean, an option that will output a boxplot of NSub number of randomly selected samples, with a default value of TRUE.}

\item{NSub}{integer, an option that specifies the number of samples to be randomly selected for the boxplot, with a default value of 10.}
}
\value{
One SummerizedExperiment class object contains both expression and metadata related to each sample and feature.
}
\description{
This function will extract data from GEO, then list out available information given in GEO, such as featureData and phenoData.
}
\examples{

GEOExtract(GEOAccession = "GSE48762", GSEMatrix =TRUE, getGPL = TRUE,  plots = TRUE, NSub = 20)

}
