% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOExtract.R
\name{GeoExtract}
\alias{GeoExtract}
\title{Process microarray data from GEO.}
\usage{
GeoExtract(GEOAccession, getGPL = TRUE, plots = TRUE, NSub = 10)
}
\arguments{
\item{GEOAccession}{character containing GEO accession}

\item{getGPL}{boolean to whether extract GPL information or not}

\item{plots}{boolean to whether plot boxplots of random chosen sample or not}

\item{NSub}{integer to number of random samples should be plotted.}
}
\value{
One SummerizedExperiment object contains both expression and metadata related to each sample.
}
\description{
This function will extract data from GEO,
then list out available information given in GEO,
such as featureData and phenoData.
We will look into featureData and phenoData before proceeding for further processing.
}
\examples{

GeoExtract(GEOAccession = "GSE48762", getGPL = TRUE,  plots = TRUE, NSub = 20)

}
