% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileNorm.R
\name{QuantileNorm}
\alias{QuantileNorm}
\title{Quantile normalization on a given eSet data.}
\usage{
QuantileNorm(GEOdata, plots = TRUE, NSub = 10)
}
\arguments{
\item{GEOdata}{either a matrix or SummerizedExperiment object contains expression.}

\item{plots}{boolean to whether plot boxplots of random chosen sample or not}

\item{NSub}{integer to number of random samples should be plotted.}
}
\value{
Either a matrix or SummerizedExperiment object contains both quantile normalized expression and metadata related to each sample.
}
\description{
This function will quantile normalize expression from a SummerizedExperiment object.
Then produce a boxplots of NSub number of randomly selected samples.
}
\examples{

QuantileNorm(GEOData = GSE48762, plots = TRUE, NSub = 10)

}
